+++ packages/frontend/src/di.ts
@@ -6,12 +6,14 @@
 import type { InjectionKey, Ref, ComputedRef } from 'vue';
 import type { PageMetadata } from '@/page.js';
 import type { Router } from '@/router.js';
+import type { NormalizedHotkeyCommand } from '@/utility/hotkey.js';
 
 export const DI = {
 	routerCurrentDepth: Symbol() as InjectionKey<number>,
 	router: Symbol() as InjectionKey<Router>,
 	mock: Symbol() as InjectionKey<boolean>,
 	pageMetadata: Symbol() as InjectionKey<Ref<PageMetadata | null>>,
+	pageCommands: Symbol() as InjectionKey<Ref<NormalizedHotkeyCommand[]>>,
 	viewId: Symbol() as InjectionKey<string>,
 	currentStickyTop: Symbol() as InjectionKey<Ref<number>>,
 	currentStickyBottom: Symbol() as InjectionKey<Ref<number>>,