+++ packages/frontend/src/router.definition.ts
@@ -117,6 +117,10 @@ export const ROUTE_DEF = [{
 		name: 'preferences',
 		component: page(() => import('@/pages/settings/preferences.vue')),
 	}, {
+		path: '/preferences/hotkeys',
+		name: 'preferences-hotkeys',
+		component: page(() => import('@/pages/settings/preferences/hotkeys.vue')),
+	}, {
 		path: '/theme/install',
 		name: 'theme',
 		component: page(() => import('@/pages/settings/theme.install.vue')),