+++ packages/frontend/src/ui/minimum.vue
@@ -16,13 +16,15 @@ import { computed, provide, ref } from 'vue';
 import { instanceName } from '@@/js/config.js';
 import XCommon from './_common_/common.vue';
 import type { PageMetadata } from '@/page.js';
-import { provideMetadataReceiver, provideReactiveMetadata } from '@/page.js';
+import type { NormalizedHotkeyCommand } from '@/utility/hotkey.js';
+import { provideMetadataReceiver, providePageCommandReceiver, provideReactiveMetadata, provideReactivePageCommands } from '@/page.js';
 import { mainRouter } from '@/router.js';
 import { DI } from '@/di.js';
 
 const isRoot = computed(() => mainRouter.currentRoute.value.name === 'index');
 
 const pageMetadata = ref<null | PageMetadata>(null);
+const pageCommands = ref<NormalizedHotkeyCommand[]>([]);
 
 provide(DI.router, mainRouter);
 provideMetadataReceiver((metadataGetter) => {
@@ -36,7 +38,11 @@ provideMetadataReceiver((metadataGetter) => {
 		}
 	}
 });
+providePageCommandReceiver((commandsGetter) => {
+	pageCommands.value = commandsGetter();
+});
 provideReactiveMetadata(pageMetadata);
+provideReactivePageCommands(pageCommands);
 </script>
 
 <style lang="scss" module>